feat: refactor and setup followship
1. change file names
2. add basic function and routers
