--- controllers/followsController.js
@@ -1,15 +1,0 @@
-const { urlencoded } = require("express");
-
-//for test only
-const helpers = require("../_helpers.js");
-const getTestUser = function (req) {
-  if (process.env.NODE_ENV === "test") {
-    return helpers.getUser(req);
-  } else {
-    return req.user;
-  }
-};
-
-const followsController = {};
-
-module.exports = followsController;
